feat: extract embeddings download
