+++ lib/searchReleaseDocs.js
@@ -1,5 +1,5 @@
 import { loadChunks, searchEmbeddings } from './embeddings.js'
-import { downloadEmbeddingsById } from './searchMarkdownDocs.js'
+import { downloadEmbeddingsById } from './downloadEmbeddings.js'
 
 const offline = process.argv.includes('--offline') || process.env.CDS_MCP_OFFLINE === 'true'
